tts_xunfei/requirements.txt
Flask==2.0.2
Werkzeug==2.0.1
websocket-client==1.2.3
Jinja2==3.1.2
